mini-rag-backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sentence-transformers>=3.0.1
faiss-cpu>=1.9.0.post1
groq>=0.9.0
numpy>=1.26.4
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
